Included needed contracts, mostly compiling

+++ wasm-client/contracts/interface/IDisputeResolutionLayer.sol
@@ -1,5 +1,5 @@
 pragma solidity ^0.4.18;
 
 interface IDisputeResolutionLayer {
-    function status(bytes32 id) external view returns (uint); //returns State enum
+    function status(bytes32 id) external view returns (uint8); //returns State enum
 }
